Rewrite test to work with HTTPS.
