+++ stripe-core/build.gradle
@@ -23,6 +23,7 @@ dependencies {
 
     ksp libs.daggerCompiler
 
+    testImplementation project(':network-testing')
     testImplementation testLibs.androidx.archCore
     testImplementation testLibs.androidx.core
     testImplementation testLibs.androidx.fragment
